mouse: Add support for getting and setting the scrollmode
Support getting and setting the scrollmode using the 'scrollmode' file.
When setting the scrollmode there are a few options. All have to be specified as
integers and are explained below.
The first integer determines the mode. The basic modes are, 1, 2, 3 and 4 for
respectively setting the scrollmode to immediate freespin, immediate
click-to-click, freespin when the mouse gets moved and click-to-click when the
mouse is moved. The other modes are 5, for automatic. 7 for switching to one
of the modes using a button. And 8 for toggling the mode using a single button.

After the number fo the mode, the duration of the setting can be set. 0 for
temporarily and 1 for default after power on. This setting is optional when
setting the mode to 1, 2, 3 of 4 and defaults to temporarily.

For the automatic mode, 5, one or two extra options must be specified. If the
second isn't specified it default to the first. This number determines the
number of needed clicks a second to switch from click-to-click to freespin.
The first number specifies the number of 'up' clicks, the second the number of
'down' clicks.

When setting the mode to two indendend switch buttons, 7, the two button
numbers have to be specified. The first one sets the button to use to switch to
freespin and the second the button to switch to click-to-click.

When setting the mode to toggle using a button only one option has to be
specified. This has to be the number of the toggle button.

Button numbers:
  1: left button (can't be used)
  2: right button (can't be used)
  3: middle (wheel) button
  4: rear thumb button
  5: front thumb button
  6: find button
  7: wheel left tilt
  8: wheel right tilt
  9: side wheel forward
 11: side wheel backward
 13: side wheel pressed