Add a debug tool, lg-debug.
Using this tool it is possible to send and receive raw commands to a device.
The device should be specified using its hidraw entry (in dev). Commands should
be entered with there hex codes, seperated with a space. For example
10 01 81 0d 00 00 00 to read the battery level. This is also how the received
bytes are getting displayed, for example 10 01 81 0d 3c b4 30 as the response
to the battery level request.